pyproject.toml
[tool.poetry]
name = "heifmgur"
version = "0.1.0"
description = "Image upload service API in Django leveraging async and HEIF"
authors = ["mbrav <mbrav@protonmail.com>"]
license = "GNU v3"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0"
Pillow = "^8.4.0"
djangorestframework = "^3.13.1"
Wand = "^0.6.7"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
